Add TUNABLES to control IFUNC selection
The current IFUNC selection is based on microbenchmarks in glibc.  It
should give the best performance for most workloads.  But other choices
may have better performance for a particular workload or on the hardware
which wasn't available at the selection was made.  The environment
variable, GLIBC_IFUNC=-xxx,yyy,-zzz...., can be used to enable CPU/ARCH
feature yyy, disable CPU/ARCH feature yyy and zzz, where the feature
name is case-sensitive and has to match the ones in cpu-features.h.  It
can be used by glibc developers to override the IFUNC selection to tune
for a new processor or improve performance for a particular workload.
It isn't intended for normal end users.

NOTE: the IFUNC selection may change over time.  Please check all
multiarch implementations when experimenting.

	* sysdeps/unix/sysv/linux/x86/dl-sysdep.c: New file.
	* sysdeps/x86/cpu-tunables.c: Likewise.
	* sysdeps/x86/dl-tunables.list: Likewise.
	* sysdeps/x86/cpu-features.c (TUNABLE_NAMESPACE): New.
	(DL_TUNABLE_CALLBACK (set_ifunc)): Likewise.
	Include <elf/dl-tunables.h> for TUNABLES is on.
	Include <string.h> and <unistd.h> if TUNABLES is off.
	(__environ): New.
	(_dl_x86_set_ifunc): Likewise.
	(init_cpu_features): Use TUNABLE_SET_VAL_WITH_CALLBACK if
	TUNABLES is on.  Call _dl_x86_set_ifunc for GLIBC_IFUNC= if
	TUNABLES is off.
	* sysdeps/x86/cpu-features.h (DEFAULT_MEMCMP): New.